# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LastOpenedRequest": {
        "properties": {
          "diagram_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagram Id"
          }
        },
        "title": "LastOpenedRequest",
        "type": "object"
      },
...
      "SaveChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "maxItems": 100,
            "title": "Messages",
            "type": "array"
          }
        },
        "title": "SaveChatRequest",
        "type": "object"
      },
      "SaveDiagramRequest": {
        "properties": {
          "title": {
            "default": "未命名架構圖",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "xml_data": {
            "maxLength": 2097152,
            "minLength": 1,
            "title": "Xml Data",
            "type": "string"
          }
        },
        "required": [
          "xml_data"
        ],
        "title": "SaveDiagramRequest",
        "type": "object"
      },
      "ShareDiagramRequest": {
        "properties": {
          "user_ids": {
            "items": {
              "type": "integer"
            },
            "title": "User Ids",
            "type": "array"
          }
        },
        "required": [
          "user_ids"
        ],
        "title": "ShareDiagramRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Cloud-360 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/collab/diagrams": {
      "get": {
        "operationId": "list_my_diagrams_api_collab_diagrams_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Diagrams"
      },
      "post": {
        "operationId": "create_diagram_api_collab_diagrams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDiagramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Diagram"
      }
    },
    "/api/collab/diagrams/{diagram_id}": {
      "delete": {
        "description": "刪除架構圖（僅擁有者）。評核／聊天經 CASCADE 一併清除；分享關聯先手動移除。",
        "operationId": "delete_diagram_api_collab_diagrams__diagram_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "diagram_id",
            "required": true,
            "schema": {
              "title": "Diagram Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Diagram"
      },
      "get": {
        "operationId": "get_diagram_api_collab_diagrams__diagram_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "diagram_id",
            "required": true,
            "schema": {
              "title": "Diagram Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Diagram"
      },
      "put": {
        "operationId": "update_diagram_api_collab_diagrams__diagram_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "diagram_id",
            "required": true,
            "schema": {
              "title": "Diagram Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDiagramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Diagram"
      }
    },
    "/api/collab/diagrams/{diagram_id}/chat": {
      "delete": {
        "description": "清空該使用者在此架構圖的對話；不刪除圖表 XML。",
        "operationId": "clear_diagram_chat_api_collab_diagrams__diagram_id__chat_delete",
        "parameters": [
          {
            "in": "path",
            "name": "diagram_id",
            "required": true,
            "schema": {
              "title": "Diagram Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear Diagram Chat"
      },
      "get": {
        "operationId": "get_diagram_chat_api_collab_diagrams__diagram_id__chat_get",
        "parameters": [
          {
            "in": "path",
            "name": "diagram_id",
            "required": true,
            "schema": {
              "title": "Diagram Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Diagram Chat"
      },
      "put": {
        "operationId": "save_diagram_chat_api_collab_diagrams__diagram_id__chat_put",
        "parameters": [
          {
            "in": "path",
            "name": "diagram_id",
            "required": true,
            "schema": {
              "title": "Diagram Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Diagram Chat"
      }
    },
    "/api/collab/diagrams/{diagram_id}/share": {
      "post": {
        "operationId": "share_diagram_api_collab_diagrams__diagram_id__share_post",
        "parameters": [
          {
            "in": "path",
            "name": "diagram_id",
            "required": true,
            "schema": {
              "title": "Diagram Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareDiagramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Share Diagram"
      }
    },
    "/api/collab/users": {
      "get": {
        "description": "分享對象列表：需架構圖編輯權。",
        "operationId": "get_users_api_collab_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Users"
      }
    },
    "/api/collab/workspace/bootstrap": {
      "get": {
        "description": "A4：進入工作區時一次取得上次開啟圖 + XML + 聊天。\n僅檢視／審核時只還原「被分享」且仍可見的圖。",
        "operationId": "workspace_bootstrap_api_collab_workspace_bootstrap_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Workspace Bootstrap"
      }
    },
    "/api/collab/workspace/last-opened": {
      "put": {
        "description": "更新使用者上次開啟的架構圖。",
        "operationId": "set_last_opened_api_collab_workspace_last_opened_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LastOpenedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Last Opened"
      }
    }
  }
}